finish refactoring brax v1
