+++ qdax/core/emitters/dpg_emitter.py
@@ -29,7 +29,7 @@ from qdax.custom_types import (
     RNGKey,
     StateDescriptor,
 )
-from qdax.tasks.brax.v1.base_wrappers import QDEnv
+from qdax.tasks.brax.v1.envs.base_env import QDEnv
 
 
 @dataclass
